requirements.txt

flask
flask-cors
python-dotenv
openai
langchain
langchain-openai
pymongo
pymupdf
python-docx
python-pptx
pywin32; sys_platform == 'win32'
pytest